requirements.txt
disnake==2.5.2
psycopg2==2.9.3
python-dotenv==0.20.0
bs4~=4.10.0
requests==2.28.1
pandas==1.4.3
git+https://github.com/Nekos-life/Async-nekos.life-wrapper
numpy~=1.19.0
keras
memory_profiler
matplotlib
beautifulsoup4==4.11.1
urllib3~=1.26.9
manim
pyserial~=3.5